tools/ai_image_editor/requirements.txt
gradio>=4.7.1
opencv-python>=4.8.1.78
numpy>=1.23.5
Pillow>=10.0.1
deepface>=0.0.75
tensorflow>=2.10.0
tensorflow-hub>=0.12.0
scikit-learn>=1.3.0
openai>=1.0.0
